fix compile warnings
Get rid of "unused argument" and "unused variable" compiler warnings.
It appears the arguments truly are not used, so not sure why they
are present. Rather than remove them, just tell compiler we know
they are not used. For the unused variable, this is because of
conditional compilation. Better fix would be to use same ifdef logic
to conditionally declare and initialize variable, but this is easy.