add clearError() routine
It appears that once an error is detected, it persists forever. And
a flash chip that is not hardcoded in to the library is an "error",
even if it is correctly recognized by the JEDEC info. Add a routine
to clear the error state so we can catch any future legitimate
errors.